Updating to v0.0.4.

+++ 03_dialog_resources/src/ids.rs
@@ -1,16 +1,16 @@
 //! IDs of the resources defined in "resources\example03.res".
 
-pub const ICO_MAIN: i32 = 101;
+pub const ICO_MAIN: u16 = 101;
 
-pub const DLG_MAIN:  i32 = 1000;
-pub const LBL_INPUT: i32 = 1001;
-pub const TXT_INPUT: i32 = 1002;
-pub const BTN_SHOW:  i32 = 1003;
+pub const DLG_MAIN:  u16 = 1000;
+pub const LBL_INPUT: u16 = 1001;
+pub const TXT_INPUT: u16 = 1002;
+pub const BTN_SHOW:  u16 = 1003;
 
-pub const DLG_MODAL:    i32 = 2000;
-pub const LBL_INCOMING: i32 = 2001;
-pub const TXT_INCOMING: i32 = 2002;
-pub const LBL_RETURN:   i32 = 2003;
-pub const TXT_RETURN:   i32 = 2004;
-pub const BTN_OK:       i32 = 2005;
-pub const BTN_CANCEL:   i32 = 2006;
+pub const DLG_MODAL:    u16 = 2000;
+pub const LBL_INCOMING: u16 = 2001;
+pub const TXT_INCOMING: u16 = 2002;
+pub const LBL_RETURN:   u16 = 2003;
+pub const TXT_RETURN:   u16 = 2004;
+pub const BTN_OK:       u16 = 2005;
+pub const BTN_CANCEL:   u16 = 2006;
